Adds provider logo to image info tab
