+++ src/styles/photodetails.scss
@@ -233,3 +233,10 @@ label {
   font-size: 0.75rem;
   font-style: italic;
 }
+
+.provider-logo {
+  margin-top: 8px;
+  margin-bottom: 8px;
+  height: 64px;
+  display: block;
+}
